Allow interface types in addition to views.
This is the large functional change. Prior to this commit interfaces were not allowed as they did not have android.view.View as their super type. This allows injecting views into interface types such as Checkable or any interface types your define yourself. The compiler will still complain if trying to inject a view into something that is neither a view subclass or an interface.