Scope symbol IDs per element package name
We need to scope that somehow in order to prevent clashes. This can be observed in a multiple modules project where butterknife is used in modules that depend on other modules that also use Butterknife.
Since IDs can repeat across modules, we need to scope them according to the package name of the current element being evaluated.