Parse manifest to determine package name.
We don't want to use the applicationId to find the location of the
R.java file because the R.java file's location is based on the package
name not the applicationId.

The applicationId can change based on the variant, but the R.java file's
location will not change.

Added some tests to make sure this works based on the FixturesTest in
SQLDelight.

Fixes #1177