Use fake ID values for R2
These values are meaningless. We only need them to be unique which is trivially achievable with a counter.

This change is required, however, because AGP 3.6 will stop generating ID values in library projects. All values are set to 0 which obviously will prevent Butter Knife from correctly performing the reverse mapping from R2 value back to name.